30/requirement.txt
opencv-python
numpy
mtcnn
Pillow